Разбивает утилитарные скрипты на модули
